Factor out no-<method>-set-state rule code to a utility
